# backend/docs/swagger.yaml
basePath: /
definitions:
  auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_expires_at:
        type: string
      cookie_strategy:
        type: string
      refresh_expires_at:
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.LoginSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/auth.LoginResponse'
      message:
        example: login successful
        type: string
      request_id:
        example: 12cd4404-a385-47fb-bc1a-eea247de4516
        type: string
      success:
        example: true
        type: boolean
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 120
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  auth.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  auth.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      updated_at:
        type: string
    type: object
  auth.UserSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/auth.UserResponse'
      message:
        example: profile fetched
        type: string
      request_id:
        example: 12cd4404-a385-47fb-bc1a-eea247de4516
        type: string
      success:
        example: true
        type: boolean
    type: object
  httpx.EmptySuccessResponse:
    properties:
      message:
        example: logout successful
        type: string
      request_id:
        example: 12cd4404-a385-47fb-bc1a-eea247de4516
        type: string
      success:
        example: true
        type: boolean
    type: object
  httpx.ErrorResponse:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        example: invalid request body
        type: string
      request_id:
        example: 12cd4404-a385-47fb-bc1a-eea247de4516
        type: string
      success:
        example: false
        type: boolean
    type: object
...
info:
  contact: {}
  description: Dokumentasi API backend TO OSN.
  title: TO OSN Backend API
  version: "1.0"
paths:
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Membuat token reset password untuk email yang terdaftar dan mengirim
        link reset melalui email. Response selalu generik agar tidak membocorkan status
        registrasi email.
      parameters:
      - description: Forgot password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.EmptySuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Request email reset password
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Melakukan autentikasi user, membuat session, lalu menulis access
        token dan refresh token ke cookie HTTP-only.
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Mencabut session aktif berdasarkan `session_id` dari access token
        lalu menghapus cookie auth.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.EmptySuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout dari device saat ini
      tags:
      - Auth
  /api/v1/auth/logout-all:
    post:
      description: Mencabut seluruh session milik user yang sedang login lalu menghapus
        cookie auth saat ini.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.EmptySuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout semua device
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      description: Membuat access token dan refresh token baru menggunakan cookie
        `refresh_token` yang masih valid.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Refresh session login
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun user baru menggunakan email, nama lengkap, dan password.
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.UserSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Register user baru
      tags:
      - Auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Mengganti password menggunakan token reset password yang masih
        valid, lalu mencabut seluruh session login user.
      parameters:
      - description: Reset password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.EmptySuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Reset password
      tags:
      - Auth
...
  /api/v1/me:
    get:
      description: Mengembalikan profil user berdasarkan subject di access token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil profil user login
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"